+++ example/lib/enumerations.dart
@@ -6,7 +6,8 @@ enum CalendarView {
   day3("Three days", Icons.view_column),
   day3Draggable("Three days - Draggable events", Icons.view_column),
   month("Month", Icons.calendar_month),
-  multi_column("Multi columns planner", Icons.view_column_outlined),
+  multi_column2("Multi columns 1", Icons.view_column_outlined),
+  multi_column1("Multi columns 2", Icons.view_column_outlined),
   day7("Seven days (web or tablet)", Icons.calendar_view_week);
 
   const CalendarView(