+++ example/lib/extension.dart
@@ -12,6 +12,21 @@ extension ColorExtension on Color {
     var newSaturation = hsl.saturation > 0.1 ? 0.48 : hsl.saturation;
     return hsl.withSaturation(newSaturation).withLightness(0.20).toColor();
   }
+
+  Color darken([double amount = .1]) {
+    assert(amount >= 0 && amount <= 1);
+    final hsl = HSLColor.fromColor(this);
+    final hslDark = hsl.withLightness((hsl.lightness - amount).clamp(0.0, 1.0));
+    return hslDark.toColor();
+  }
+
+  Color lighten([double amount = .1]) {
+    assert(amount >= 0 && amount <= 1);
+    final hsl = HSLColor.fromColor(this);
+    final hslLight =
+        hsl.withLightness((hsl.lightness + amount).clamp(0.0, 1.0));
+    return hslLight.toColor();
+  }
 }
 
 extension DateTimeExtensions on DateTime {